Extractor builds without any reference to syntax (as it will be helping to produce this!).
